Feat: Added a PaintPixel action. Added MouseBinding support to the BindingManager.
Added a PaintPixel action to the enum class Action and added MouseBinding support to the BindingManager to allow for user interaction through the mouse. The user can paint pixels onto the canvas with left click. CommandManager.execute function now requires many parameters. A rework of the command system is planned.