Feat: Added a basic ILayout class that all layouts will inherit from.
Added a basic ILayout class that all layouts will inherit from. This gives some base functionality for UI layouts and provides update and render methods for them. The Canvas SurfaceView no longer has a StatusBar property, instead a Canvas Layout now has a SurfaceView and StatusBar property and acts as a container for the Canvas View and UI objects.