Refactor: Moved orchestration functions from EditorController to Application class.
This commit significantly refactors the code base by shifting sever critical orchestration functions from the EditorController to the more appropriate Application class.

The main run loop for the Editor now lives in the Application class as well as the Input Manager and Render Manager. The EditorController class now acts more like a controller.